refactoring test snapshots
